Added an LED and Buzzer
Added an LED and Buzzer for pre-notification of servo rotation, also
refactored the delay between servo rotations to not use the delay()
method, but rather use millis() and save off when to execute the next
rotation. The advantage here is we can then press the button to make the
servos rotate whenever we want. If we use a 12 hour delay, for example,
no other inputs could be easily captured.