Added another servo
Added a right servo. Initially both servos would rotate clockwise. Now
the left servo rotates clockwise and the right servo rotates counter
clockwise. The reason is the food would get stuck in the PVC elbow of
the right servo, since the elbow was pointed in the opposite direction
the food was being dispensed.

I also adjusted the pauseInterval for a time lag I was noticing of
exactly 2 minutes, so in reality the food was being dispensed every 12
hrs and 2 minutes.

I have also wired up a momentary push button externally, which will
reset the timer to the 12 hour delay when pressed, which also means the
lastExecution variable doesn't need a initialization offset.